package friendlier and minor updates
