+++ pyeafe/assembly.py
@@ -8,7 +8,7 @@ convection diffusion reaction equation:
 Integral quantities are approximated by quadrature.
 
 Usage:
-    eafe(mesh, diff, conv, reac, boundary)
+    eafe_assemble(mesh, diff, conv, reac, boundary)
 
     - mesh: mesh defining finite element space
     - diff: DOLFIN expression for diffusion
@@ -61,7 +61,7 @@ def bernoulli(r):
         return r * np.exp(-r) / (1 - np.exp(-r))
 
 
-def eafe(mesh, diff, conv=None, reac=None, boundary=None, **kwargs):
+def eafe_assemble(mesh, diff, conv=None, reac=None, boundary=None, **kwargs):
     quadrature_degree = parameters["form_compiler"]["quadrature_degree"]
     parameters["form_compiler"]["quadrature_degree"] = 2
     spatial_dim = mesh.topology().dim()
