+++ pyeafe/tests/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ["double_sin_test","double_sin_reaction_test","double_sin_3d_test"]
