--- src/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ["eafe"]
