+++ tests.py
@@ -1,0 +1,6 @@
+'''
+ Script to test the EAFE module
+'''
+
+from pyeafe import run_tests as rt
+rt.run()
